backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2
alembic
uvicorn
pwdlib[argon2]
PyJWT
stripe